requirements.txt
crewai>=0.16.1
python-dotenv>=1.0.0
requests>=2.31.0
groq>=0.4.0
langchain>=0.1.0
flask>=2.0.0
flask-cors>=3.0.10 